LPS-60972 Plugin to configure gradle-plugins-lang-builder
